arch: relocate clientgraphics out of the pure domain (model/ -> root)
model/clientgraphics imported oblikovati.org/renderer and oblikovati.org/scene,
the one concrete violation of the 'domain imports nothing above it' rule — a
viewport adapter misfiled under model/. In substance it builds renderer.DrawItems
from a scene camera, so it is presentation, not domain.

Move it to a top-level clientgraphics/ package (alongside renderer/ and scene/,
where depending on them is legal) and rewrite the 10 importers in addin/router,
app, and head. Pure move: the package's own files were unchanged (no self-path
import); no behavior change. After this, NO package under model/ imports
renderer or scene — the domain is GPU-free and headless-testable again, which
unblocks the domain-purity archguard (#1499).

Closes #1500
