+++ clientgraphics/colormap.go
