+++ clientgraphics/colormap_test.go
