+++ clientgraphics/decode.go
