+++ clientgraphics/fan_test.go
