+++ clientgraphics/glyph.go
