+++ clientgraphics/image_test.go
