+++ clientgraphics/object_model_test.go
