+++ clientgraphics/store.go
