+++ clientgraphics/store_test.go
