+++ clientgraphics/strip_test.go
