+++ clientgraphics/surface_test.go
