+++ head/cmd/m16shot/main.go
@@ -21,11 +21,11 @@ import (
 	"oblikovati.org/api/types"
 	"oblikovati.org/api/wire"
 	"oblikovati.org/app"
+	"oblikovati.org/clientgraphics"
 	"oblikovati.org/head/internal/native"
 	"oblikovati.org/head/ui"
 	"oblikovati.org/kernel/ops"
 	"oblikovati.org/math"
-	"oblikovati.org/model/clientgraphics"
 	"oblikovati.org/model/compdef"
 	"oblikovati.org/model/feature"
 	"oblikovati.org/model/sketch"
